:recycle: Extract files manipulation to FileService out of Files.vue and FileFolderSelectModal.vue + fix clicking on select file hitting always the first file
