+++ frontend/src/Files.vue
@@ -45,7 +45,7 @@
           hidden="hidden"
           multiple
           type="file"
-          @change.stop="(e) => uploadFiles((e.target as HTMLInputElement).files)"
+          @change.stop="(e) => uploadFiles((e.target as HTMLInputElement).files, false)"
         />
         <label
           class="btn btn-primary btn-sm me-2"
@@ -219,8 +219,8 @@
           </div>
           <div class="modal-body">
             <div class="text-center pb-2">
-              {{ currentUpload.progress === 100 ? 'Saving' : 'Uploading' }}
-              {{ currentUpload.name }} ...
+              {{ currentUpload?.progress === 100 ? 'Saving' : 'Uploading' }}
+              {{ currentUpload?.name }} ...
             </div>
             <div
               aria-label="Example with label"
@@ -233,12 +233,12 @@
               <div
                 :class="{
                   'progress-bar-striped  progress-bar-animated':
-                    currentUpload.progress === 100,
+                    currentUpload?.progress === 100,
                 }"
-                :style="{ width: currentUpload.progress + '%' }"
+                :style="{ width: currentUpload?.progress + '%' }"
                 class="progress-bar"
               >
-                {{ currentUpload.progress }}%
+                {{ currentUpload?.progress }}%
               </div>
             </div>
           </div>
@@ -247,7 +247,7 @@
               class="btn btn-danger"
               data-bs-dismiss="modal"
               type="button"
-              @click.stop="currentUploadAbortController?.abort()"
+              @click.stop="abortController?.abort()"
             >
               Cancel
             </button>
@@ -262,43 +262,27 @@
 
 <script setup lang="ts">
 import router from './router';
-import AuthService from './services/AuthService';
-import { computed, nextTick, ref, watchEffect, onBeforeUnmount } from 'vue';
+import { computed, nextTick, onBeforeUnmount, onMounted, ref, watchEffect } from 'vue';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import utils from './utils';
-import axios, { AxiosProgressEvent } from 'axios';
-import { Modal } from 'bootstrap';
 import TransferHelpModal from './components/TransferHelpModal.vue';
-
-interface FileItem {
-  name: string;
-  path: string;
-  size: number;
-  type: string;
-}
+import { FileItem, FileService, UploadProgress, ValidationResult } from '@/services/FileService.ts';
+import { Modal } from 'bootstrap';
 
 interface Props {
   paths: string | string[];
 }
 
-interface UploadProgress {
-  name: string;
-  progress: number;
-}
-
-interface ValidationResult {
-  valid: boolean;
-  message?: string | null;
-}
 
 const props = defineProps<Props>();
 const files = ref<FileItem[]>([]);
 const loading = ref(false);
-const currentUpload = ref<UploadProgress>({ name: '', progress: 0 });
+const currentUpload = ref<UploadProgress | null>(null);
+const uploadModal = ref<Modal>();
 const sortKey = ref<'name' | 'size'>('size');
 const sortDirection = ref<1 | -1>(1);
-const fetchAbortController = ref(new AbortController());
-const currentUploadAbortController = ref<AbortController | null>(null);
+const fetchAbortController = ref<AbortController>(new AbortController());
+const abortController = ref<AbortController>(new AbortController());
 
 const renamingFileKey = ref<string | null>(null);
 const renamingFileName = ref<string>('');
@@ -306,14 +290,6 @@ const renamingOngoing = ref(false);
 const renamingInput = ref<HTMLInputElement[] | null>(null);
 const labelRefs: Record<string, any> = {};
 const renamingMessage = ref<string | null>(null);
-
-//TODO: Get these constants from config
-const HARD_UPLOAD_SIZE_LIMIT = 30; // in GBs;
-const MAX_UPLOAD_SIZE = 1024; // in MBs;
-const MAX_UPLOAD_FILE_COUNT = 1000;
-const KB = 1024;
-const MB = KB * 1024;
-const GB = MB * 1024;
 
 const currentPath = computed(() => props.paths === '' ? [''] : ['', ...(Array.isArray(props.paths) ? props.paths : [props.paths])]);
 const sortedFiles = computed(() => files.value?.sort((a, b) => // sort on type before name
@@ -326,17 +302,8 @@ const showToast = ref(false);
 
 onBeforeUnmount(() => {
   fetchAbortController.value.abort();
-  currentUploadAbortController.value?.abort();
+  abortController.value?.abort();
   loading.value = false;
-
-  // Clean up Bootstrap modal
-  const modalElement = document.getElementById('uploadModal');
-  if (modalElement) {
-    const modalInstance = Modal.getInstance(modalElement);
-    if (modalInstance) {
-      modalInstance.dispose();
-    }
-  }
 });
 
 const triggerToast = (message: string, duration = 3000) => {
@@ -355,36 +322,14 @@ const flipSort = (key: 'name' | 'size') => {
   }
 };
 
-const fetchFiles = async () => {
-  if (!AuthService.isAuthenticated()) return;
-
-  loading.value = true;
-  fetchAbortController.value = new AbortController();
-  const pathArray = Array.isArray(props.paths) ? props.paths : (props.paths ? [props.paths] : []);
-  const path = props.paths && props.paths !== '' ? pathArray.join('/') : '';
-
-  try {
-    const response = await axios.post<FileItem[]>(
-      '/api/files/user/query',
-      { path },
-      { signal: fetchAbortController.value.signal },
-    );
-    files.value = response.data.map(file => ({
-      ...file,
-      path: file.path === 'user' ? '' : file.path.replace(/^user\/?/, ''),
-    }));
-  } catch (error: any) {
-    // Ignore abort errors - these are intentional cancellations
-    if (error.code === 'ERR_CANCELED' || error.name === 'AbortError') {
-      return;
-    }
-    console.error('Failed to fetch files:', error);
-  } finally {
-    loading.value = false;
-  }
-};
+const fetchFiles = async() => {
+  files.value = await FileService.fetchFiles(currentPath.value.slice(1), loading);
+}
 
 watchEffect(fetchFiles);
+
+onMounted(() => uploadModal.value = new Modal('#uploadModal'))
+watchEffect(() => currentUpload.value ? uploadModal.value?.show() : uploadModal.value?.hide())
 
 const navigate = async (path: string, name: string) => {
   path = path.split('/').map(encodeURIComponent).filter(Boolean).join('/');
@@ -401,66 +346,27 @@ const navigate = async (path: string, name: string) => {
 };
 
 
-const downloadFile = (file: FileItem) => {
-  axios.post('/api/files/user/download', {
-    path: file.path,
-    fileName: file.name,
-  }, {
-    responseType: 'blob',
-  }).then((response) => {
-    const a = document.createElement('a');
-    a.href = URL.createObjectURL(response.data);
-    a.setAttribute('download', file.name);
-    a.click();
-  }).catch((error) => {
-    console.error('Download failed:', error);
-  });
-};
-
-const downloadFileList = (file: FileItem) => {
-  const downloadPath = ['/filelist', file.path, file.name].filter(Boolean).join('/');
-  axios({ url: downloadPath, responseType: 'blob' }).then((response) => {
-    const a = document.createElement('a');
-    a.href = URL.createObjectURL(response.data);
-    a.setAttribute('download', file.name + '.tsv');
-    a.click();
-  });
+const uploadFiles = async (uploads: FileList | null, isFolderUpload: boolean) => {
+  await FileService.uploadFiles(uploads, isFolderUpload, currentPath.value, files.value, currentUpload, abortController);
+  await fetchFiles();
 };
 
 const deleteFile = async (file: FileItem) => {
-  if (!await utils.confirm('Delete File', `Do you want to delete ${file.name}?`, { okayLabel: 'Delete' })) return;
-  const body = {
-    path: file.path,
-    fileName: file.name,
-  };
-  try {
-    loading.value = true;
-    await axios.post('/api/files/user/delete', body);
-    await fetchFiles(); // refresh file list
+  if (await FileService.deleteFile(file, loading)) {
     triggerToast(`✅ "${file.name}" deleted successfully.`);
-  } catch (error) {
-    console.error('Failed to delete file:', error);
+    await fetchFiles();
   }
 };
 
 const renameFile = async (file: FileItem, newName: string): Promise<boolean> => {
-  if (file.name === newName) return true;
-  const body = {
-    path: file.path,
-    originalName: file.name,
-    newName,
-  };
-
-  try {
-    await axios.post('/api/files/user/rename', body, { responseType: 'json' });
-    await fetchFiles(); // refresh file list
-    return true;
-  } catch (error: any) {
-    console.error('Failed to rename file:', error);
-    renamingMessage.value = error.response?.data?.log?.message || 'Failed to rename file';
-    return false;
-  }
+  await FileService.renameFile(file, newName, renamingMessage);
+  await fetchFiles();
+  return true;
 };
+
+const downloadFile = (file: FileItem) => FileService.downloadFile(file);
+const downloadFileList = (file: FileItem) => FileService.downloadFileList(file);
+
 
 const startRenaming = (file: FileItem) => {
   const fileKey = file.path + '/' + file.name;
@@ -542,157 +448,6 @@ const checkExtension = async (oldName: string, newName: string): Promise<boolean
     );
   }
   return true;
-};
-
-const uploadFile = (file: File) => {
-  currentUploadAbortController.value = new AbortController();
-
-  const formData = new FormData();
-
-  // Attach the file
-  formData.append('files', file);
-
-  // Attach JSON as a string under key 'filePath'
-  formData.append(
-    'filePath',
-    currentPath.value
-      .slice(1)
-      .map((a) => decodeURIComponent(a))
-      .join('/'),
-  );
-
-  return axios.post(`/api/files/user/upload`, formData, {
-    signal: currentUploadAbortController.value.signal,
-    headers: {
-      'Content-Type': 'multipart/form-data',
-    },
-
-    onUploadProgress: ((progressEvent: AxiosProgressEvent) => {
-      if (progressEvent.total) currentUpload.value.progress = Math.round((progressEvent.loaded / progressEvent.total) * 100);
-    }),
-  });
-};
-
-
-const uploadFiles = async (uploads: FileList | null, isFolderUpload?: boolean) => {
-  if (!uploads) return false;
-
-  let filesToUpload: File[] = Array.from(uploads);
-
-  if (filesToUpload.find((file) => file.size / GB > HARD_UPLOAD_SIZE_LIMIT)) {
-    await utils.confirm(
-      `Upload error`,
-      `You are trying to upload at least one file larger than ${HARD_UPLOAD_SIZE_LIMIT} GB.<br>This is not supported by this web interface.<br>Please use FTP/Aspera for such large files transfer.`,
-      { showCancel: false },
-    );
-    return;
-  }
-
-  // warn when files are large or more than a threshold
-  const largeFiles = filesToUpload.filter(
-    (file) => file.size / MB > MAX_UPLOAD_SIZE,
-  );
-  if (largeFiles.length > 0 || filesToUpload.length > MAX_UPLOAD_FILE_COUNT) {
-    const proceed = await utils.confirm(
-      `Upload warning`,
-      `For uploading files larger than ${MAX_UPLOAD_SIZE} MB or more than ${MAX_UPLOAD_FILE_COUNT} files, using FTP/Aspera is recommended. Do you still want to continue?`,
-      { okayLabel: `Yes`, cancelLabel: 'No, cancel' },
-    );
-    if (!proceed) return;
-  }
-
-  // warn overwrite
-  const fileNames = files.value.map((f) => f.name);
-  const overlap = filesToUpload
-    .map((file) =>
-      isFolderUpload
-        ? file?.webkitRelativePath.substring(
-          0,
-          file?.webkitRelativePath.indexOf('/'),
-        )
-        : file.name,
-    )
-    .filter((name) => fileNames.includes(name));
-
-  if (overlap.length > 0) {
-    const overlapString =
-      overlap.length === 1
-        ? overlap[0] + '?'
-        : overlap.length + ' files? (' + overlap.join(', ') + ')';
-    const proceed = await utils.confirm(
-      `Overwrite files?`,
-      isFolderUpload
-        ? 'This may overwrite existing files in the folder. Do you want to go ahead?'
-        : `Do you want to overwrite ${overlapString}`,
-      { okayLabel: `Yes, overwrite`, cancelLabel: 'No, cancel' },
-    );
-    if (!proceed) return;
-  }
-
-  // filter hidden files
-  const hiddenFiles = filesToUpload.filter((file) => file.name.startsWith('.'));
-  if (hiddenFiles.length > 0) {
-    // Count filenames to avoid repeating several times the same name
-    const counter = hiddenFiles.map(file => file.name)
-      .reduce(
-        (acc, name) => {
-          acc[name] ? acc[name]++ : acc[name] = 1;
-          return acc;
-        },
-        {} as Record<string, number>,
-      );
-    // Sort by how many time file names appear, and then by their name
-    const fileNamesCounts = Object.entries(counter).sort(
-      ([nameA, countA], [nameB, countB]) =>
-        countB - countA || nameA.localeCompare(nameB),
-    );
-    // Only display the top 3 kind of hidden filenames
-    const examples =
-      fileNamesCounts.length > 3
-        ? fileNamesCounts.slice(0, 3)
-        : fileNamesCounts;
-
-    const introDisplay =
-      hiddenFiles.length > 1
-        ? `Do you want to include these <b>${hiddenFiles.length}</b> hidden files?`
-        : `Do you want to include this hidden file?`;
-    let examplesDisplay = examples
-      .map(
-        ([name, count]) =>
-          `<li>${count > 1 ? `<b>${count} x </b> ` : ''} ${name}</li>`,
-      )
-      .join('');
-    if (examples.length !== fileNamesCounts.length)
-      examplesDisplay += '<li>...</li>';
-
-    const filter = await utils.confirm(
-      `Upload hidden files?`,
-      `${introDisplay} <ul>${examplesDisplay}</ul>`,
-      {
-        okayLabel: `Skip hidden files (${filesToUpload.length - hiddenFiles.length} files)`,
-        cancelLabel: `Include all files (${filesToUpload.length} files)`,
-        level: 'primary',
-      },
-    );
-    if (filter)
-      filesToUpload = filesToUpload.filter(
-        (file) => !file.name.startsWith('.'),
-      );
-  }
-
-  const uploadModal = new Modal('#uploadModal');
-  uploadModal.show();
-
-  for (let i = 0; i < filesToUpload.length; i++) {
-    const file = filesToUpload[i];
-    currentUpload.value.name = file.name;
-    await uploadFile(file)
-      .catch((e) => {
-        if (e?.message === 'canceled') i = filesToUpload.length; // Cancel everything when hitting cancel
-      });
-  }
-  await fetchFiles();
-  return false;
 };
 </script>
