+++ frontend/src/components/FileFolderSelectModal.vue
@@ -3,18 +3,19 @@
 import { ComponentPublicInstance, computed, getCurrentInstance, inject, onMounted, ref } from 'vue';
 import FileTree from './FileTree.vue';
 import { Modal } from 'bootstrap';
-import axios, { AxiosProgressEvent } from 'axios';
+import axios from 'axios';
 import { useRoute } from 'vue-router';
 import { PageTab } from '@/models/PageTab.model.ts';
 import type { Node } from '@/models/FileTreeNode.model.ts';
-import utils, { Class } from '@/utils.ts';
+import { Class } from '@/utils.ts';
+import { FileService, UploadProgress } from '@/services/FileService.ts';
 
 const props = defineProps<{
   file: PageTab.File | PageTab.Attribute,
   class?: Class,
   allowFolder?: boolean,
   isFileList?: boolean,
-  row?: PageTab.File
+  row?: PageTab.Section
 }>();
 const emits = defineEmits<{ select: [PageTab.File | PageTab.Attribute] }>();
 const thisFile = ref(props.file);
@@ -23,13 +24,15 @@ if (isFileList.value === undefined && (thisFile.value as PageTab.Attribute)?.nam
   isFileList.value = true;
 }
 const allowFolder = !isFileList.value;
-const axiosAbortController = new AbortController();
-const currentUpload = ref({ name: '', progress: 0 });
-const showProgressbar = ref(false);
-const MAX_UPLOAD_SIZE = 1024; //1GB
+const abortController = ref<AbortController>(new AbortController());
+const currentUpload = ref<UploadProgress | null>(null);
+
+
 const errorMessage = ref('');
 const route = useRoute();
 const parentDisplayType = inject('parentDisplayType');
+
+const path = [] as string[];
 
 
 const isAFileList = (_file: PageTab.File | PageTab.Attribute): _file is PageTab.Attribute => isFileList.value;
@@ -41,55 +44,24 @@ const thisComponent = getCurrentInstance();
 const filetree = ref<ComponentPublicInstance<typeof FileTree>>();
 onMounted(() => {
   let elementById = document.getElementById('fileFolderSelectModal' + thisComponent!.uid);
-  if (elementById)
-    modal = new Modal(elementById, {
-      backdrop: 'static',
-    });
+  if (elementById) modal = new Modal(elementById, { backdrop: 'static' });
 });
 
-const uploadFile = async (file: File) => {
-  if (file.size > MAX_UPLOAD_SIZE * 1024 * 1024) { // Check if file size is greater than 1GB
-    const proceed = await utils.confirm(
-      `Upload warning`,
-      `For uploading larger than ${MAX_UPLOAD_SIZE} MB, using FTP/Aspera is recommended. Do you still want to continue?`,
-      { okayLabel: `Yes`, cancelLabel: 'No, cancel' });
-    if (!proceed) return;
-  }
-  const formData = new FormData();
-  formData.append('files', file);
-  showProgressbar.value = true;
+const uploadFile = async (file: FileList) => {
   try {
-    await axios.post(`/api/files/user/`, formData, {
-      signal: axiosAbortController.signal,
-      headers: {
-        'Content-Type': 'multipart/form-data',
-      },
-      onUploadProgress: function(progressEvent: AxiosProgressEvent) {
-        const progress = Math.round((progressEvent.loaded / progressEvent.total!) * 100);
-        currentUpload.value.progress = progress;
-
-        if (progress === 100) {
-          // Wait a bit for the backend to process the file after it's fully uploaded
-          setTimeout(() => {
-            if (modal) {
-              modal.hide(); // Hide the modal when upload is complete
-            }
-          }, 500); // Adjust the timeout as necessary
-        }
-      },
-    });
-    filetree.value?.show(); // Assuming refreshTree is a method in FileTree
+    const currentFiles = await FileService.fetchFiles(path);
+    const succeed = await FileService.uploadFiles(file, false, [], currentFiles, currentUpload, abortController);
+    if (!succeed) return;
+    loadTree(); // Assuming refreshTree is a method in FileTree
+    modal?.hide()
     //We are reusing this component for file and file list selection. File is using path in pagetab but FileList use value
     //this line is trying to set correct pageTab variable based on component usage
-    isAFileList(thisFile.value) ? thisFile.value.value = file.name : thisFile.value.path = file.name; // Update the input box with the file name
-    if (isFileList.value) {
-      await validateFileListFile(file.name);
+    isAFileList(thisFile.value) ? thisFile.value.value = file[0].name : thisFile.value.path = file[0].name; // Update the input box with the file name
+    if (isAFileList(thisFile.value)) {
+      await validateFileListFile(file[0].name);
     }
   } catch (error) {
     console.error('File upload error:', error);
-  } finally {
-    showProgressbar.value = false;
-    currentUpload.value.progress = 0;
   }
 };
 
@@ -151,7 +123,6 @@ const loadTree = () => filetree.value?.show();
     <div class="input-group input-group-sm" :class="props.class">
       <input type="text" class="form-control bg-body-secondary" v-model="fileModalModel" readonly data-bs-toggle="modal"
              :disabled="parentDisplayType==='readonly'"
-
              :data-bs-target="'#fileFolderSelectModal'+thisComponent!.uid"
              @click="loadTree()" :class="{'is-invalid': errorMessage}">
       <button class="btn btn-secondary" type="button" data-bs-toggle="modal"
@@ -184,12 +155,12 @@ const loadTree = () => filetree.value?.show();
           <h5>Upload File</h5>
           <div class="input-group">
             <div class="custom-file">
-              <input id="inputGroupFile" type="file" class="custom-file-input"
-                     @change="uploadFile(($event.target as HTMLInputElement).files![0])">
-              <label for="inputGroupFile" class="custom-file-label"> Select file to upload </label>
+              <input :id="'inputGroupFile'+thisComponent!.uid" type="file" class="custom-file-input"
+                     @change="uploadFile(($event.target as HTMLInputElement).files!)">
+              <label :for="'inputGroupFile'+thisComponent!.uid" class="custom-file-label"> Select file to upload </label>
             </div>
           </div>
-          <div v-if="showProgressbar">
+          <div v-if="currentUpload !== null">
             <div class="text-center pb-2">{{ currentUpload.progress === 100 ? 'Saving' : 'Uploading' }}
               {{ currentUpload.name }} ...
             </div>
@@ -204,7 +175,7 @@ const loadTree = () => filetree.value?.show();
         </div>
         <div class="modal-footer">
           <button type="button" :id="'fileFolderSelectModal'+thisComponent!.uid" class="btn btn-outline-secondary"
-                  data-bs-dismiss="modal" @click.stop="axiosAbortController.abort()">Cancel
+                  data-bs-dismiss="modal" @click.stop="abortController.abort()">Cancel
           </button>
         </div>
       </div>
